Register Signal Callbacks
This commit registers two signals we need to listen on.
The first signal (`buffer_line_added`) gets fired whenever a new line
is added on any of the buffers. Our task is to identify the type of
buffer where a line got added (query, server buffer, irc channel,… ) and
unhide that buffer if it matches our criteria.

The second signal (`buffer_switch`) gets fired whenever the program user
switches from one buffer to another. Our task is to identify the type
of the previous buffer (query, server buffer, irc channel,… ) and hide
that buffer from the buflist if it matches our criteria.

I will continue once I am back in the mood again .-.